fix(diffusers): fix bugs found in validation

+++ mindone/diffusers/models/unets/unet_motion_model.py
@@ -839,7 +839,7 @@ class UNetMotionModel(ModelMixin, ConfigMixin, UNet2DConditionLoadersMixin):
                     f"{self.__class__} has the config param `addition_embed_type` set to 'text_time' which requires the keyword argument `time_ids` to be passed in `added_cond_kwargs`"  # noqa: E501
                 )
             time_ids = added_cond_kwargs.get("time_ids")
-            time_embeds = self.add_time_proj(time_ids.flatten())
+            time_embeds = self.add_time_proj(time_ids.flatten()).to(text_embeds.dtype)
             time_embeds = time_embeds.reshape((text_embeds.shape[0], -1))
 
             add_embeds = ops.concat([text_embeds, time_embeds], axis=-1)
